+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## 1.0.1 (5/10/2022)
+
+- Wait, I have a changelog?
+- Added poe.ninja support for the following items:
+  - Cluster Jewels
+  - Maps
+  - Blighted Maps
+  - Blight-Ravaged Maps
+  - Unique Maps
+  - Delirium Orbs
+  - Invitations
+  - Scarabs
+- Added poe.ninja support with proper slash commands for generic price checks. Still working through the item types. Check the readme for more info.
+
 ## 1.0.0
 
 - Initial version.